+++ src/transport/campfire/campfire_handler.erl
@@ -21,7 +21,9 @@
         % bot nick in campfire room
         campfire_nick = <<>> :: binary(),
         % campfire client process pid
-        campfire_client_pid :: pid()
+        campfire_client_pid :: pid(),
+        % parser process pid
+        parser_pid :: pid()
     }).
  
 start_link() ->
@@ -33,8 +35,8 @@ init([]) ->
 handle_call(_Request, _From, State) ->
     {reply, ignored, State}.
 
-handle_cast({campfire_client, ClientPid, Login}, State) ->
-    {noreply, State#state{campfire_client_pid = ClientPid, campfire_nick = Login}};
+handle_cast({campfire_client, ClientPid, ParserPid, Login}, State) ->
+    {noreply, State#state{campfire_client_pid = ClientPid, parser_pid = ParserPid, campfire_nick = Login}};
 
 handle_cast(_Msg, State) ->
     {noreply, State}.
@@ -51,48 +53,8 @@ handle_info({incoming_message, IncomingMessage}, State) ->
                          end,
     % Get body
     {_, Body} = lists:keyfind(<<"body">>, 1, DataList),
-    % Get message body
-    Message = string:tokens(binary_to_list(Body), " \r\n"),
-    % Check this is message for Ybot or not
-    case Message of
-        [Nick] ->
-            gen_server:cast(State#state.campfire_client_pid, {send_message, "", "What?"});
-        [Nick, "hi"] ->
-            gen_server:cast(State#state.campfire_client_pid, {send_message, "", "Hello :)"});
-        [Nick, "bye"] ->    
-            gen_server:cast(State#state.campfire_client_pid, {send_message, "", "Good bue"});
-        [Nick, "history"] ->
-            % Get history
-            History = gen_server:call(ybot_history, {get_history, State#state.campfire_client_pid}),
-            % Send history
-            gen_server:cast(State#state.campfire_client_pid, {send_message, "", History});
-        [Nick, "plugins?"] ->
-            % Get plugins
-            Plugins = gen_server:call(ybot_manager, get_plugins),
-            % Send plugins label
-            gen_server:cast(State#state.campfire_client_pid, {send_message, "", "Plugins:"}),
-            % Make plugins list
-            lists:foreach(fun(Plugin) ->
-                              {_, _, Pl, _} = Plugin,
-                              gen_server:cast(State#state.campfire_client_pid, {send_message, "", "Plugin: " ++ Pl ++ "\r\n"})
-                          end, 
-                          Plugins),
-            gen_server:cast(State#state.campfire_client_pid, {send_message, "", "That's all :)"});
-        [Nick, Command | Arg] ->
-            Args = case Arg of
-                    [] ->
-                        % return empty args
-                        "";
-                    _ ->
-                        % Get command arguments
-                        ybot_utils:split_at_end(binary_to_list(Body), Command)
-                    end,
-            % Start process with supervisor which will be execute plugin and send to pid
-            ybot_actor:start_link(State#state.campfire_client_pid, "", Command, Args);
-        _ ->
-            % this is not our command
-            pass
-    end,
+    % Send message to parser
+    gen_server:cast(State#state.parser_pid, {incoming_message, State#state.campfire_client_pid, Nick, "", binary_to_list(Body)}),
     % return
     {noreply, State};