+++ Changelog.md
@@ -1,6 +1,8 @@
 # Ybot news & changelog
 
-## Ybot-0.3.1 --> In development
+## Ybot-0.3.2 --> In development
+
+## Ybot-0.3.1 --> released
 
   * Added reconnect timeout option and reconnect ability to irc client --> #33.
   * hacker_help plugin fixed --> #35
@@ -16,7 +18,8 @@
   * Skype support added.
   * Removed http PUT request supporting. Now only POST.
   * New transport - Talkerapp (http://talkerapp.com/rooms).
-  
+  * Message parsers moved to ybot_parser module.
+
 ## Ybot-0.3 --> released
 
   * New wat plugin